Rewrite message, fix Lambda+SES permissions
With these changes, I'm now able to successfully send email and receive
it forwarded back to me. The messages pass SPF, DKIM, and DMARC
validation, though for the moment the Gmail "Show original" view only
highlights DKIM as passing. I'll give it a day or two to see if it has
to do with DNS record caches waiting to time out.

It turns out we can't just pass the raw message through. SES will parse
it to ensure that it's being sent from a verified address. This is a
good thing, as it doesn't allow one to spoof email sent from another
domain. But it did require some message manipulation (which I will write
proper tests for in a bit).

Also, the SESCrudPolicy wasn't permissive enough to allow SendRawEmail
to work. It failed with:

  failed to forward message incoming/..: operation error SES:
  SendRawEmail, https response error StatusCode: 403, RequestID: ...,
  api error AccessDenied: User
  `arn:aws:sts::...:assumed-role/ses-forwarder.../ses-forwarder-Function-...'
  is not authorized to perform `ses:SendRawEmail' on resource
  `arn:aws:ses:us-east-1:...:configuration-set/ses-forwarder':
  errorString null

Replacing it with SESSendEmailPolicy after a hint from
https://stackoverflow.com/a/75023245 did the trick.