Extract updateHeadersInput, lowercase most methods
It was bothering me that the headerBuffer had so many members that only
a few of the methods used. Passing the inputs as an argument to
WriteUpdatedHeaders makes explicit where each of these input fields
apply. Higher cohesion! Lower coupling!

The only method intended for direct use by the handler is
WriteUpdateHeaders. Lowercasing the other methods should help
communicate this.