Finish headerBuffer tests, rename writeFrom
In the process of writing these tests, I renamed and refactored
writeFromHeader to writeFromAndReplyTo.

The tests helped me catch a potential bug in this refactoring whereby
two Reply-To headers would appear if a Reply-To header already existed.
writeFromAndReplyTo would emit it once, and because Reply-To was still
in keepHeaders, WriteUpdatedHeaders would emit it again. The solution
was to remove it from keepHeaders.

This is why we unit test everything as close to the logic as possible,
even after line coverage is 100% from other tests!